refactor: reorganize print format builder into components/stores structure
