+++ frappe/public/js/print_format_builder/components/Preview.vue
@@ -38,7 +38,7 @@
 </template>
 
 <script setup>
-import { useStore } from "./store";
+import { useStore } from "../stores";
 import { ref, computed, onMounted } from "vue";
 
 // mixin
