+++ frappe/public/js/print_format_builder/components/inspector/ConfigureColumns.vue
